@@ -5,7 +5,7 @@ version: 1.2.4
 homepage: https://github.com/fayaz07/progress_dialog
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
