@@ -1,13 +1,11 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/painting.dart';
 
 enum ProgressDialogType { Normal, Download }
 
 String _dialogMessage = "Loading...";
 double _progress = 0.0, _maxProgress = 100.0;
 
-Widget _customBody;
+late Widget _customBody;
 
 TextAlign _textAlign = TextAlign.left;
 Alignment _progressWidgetAlignment = Alignment.centerLeft;
@@ -16,8 +14,8 @@ Alignment _dialogAlignment = Alignment.center;
 TextDirection _direction = TextDirection.ltr;
 
 bool _isShowing = false;
-BuildContext _context, _dismissingContext;
-ProgressDialogType _progressDialogType;
+late BuildContext _context, _dismissingContext;
+late ProgressDialogType _progressDialogType;
 bool _barrierDismissible = true, _showLogs = false;
 
 TextStyle _progressTextStyle = TextStyle(
@@ -36,43 +34,43 @@ Widget _progressWidget = Image.asset(
 );
 
 /// For Auto Hide Dialog after some Duration.
-Duration _autoHide;
+late Duration _autoHide;
 
 class ProgressDialog {
-  _Body _dialog;
+  late _Body _dialog;
 
   ProgressDialog(BuildContext context,
-      {ProgressDialogType type,
-      bool isDismissible,
-      bool showLogs,
-      Duration autoHide,
-      TextDirection textDirection,
-      Widget customBody}) {
+      {ProgressDialogType type = ProgressDialogType.Normal,
+      bool? isDismissible,
+      bool showLogs = false,
+      Duration autoHide = const Duration(seconds: 2),
+      TextDirection textDirection = TextDirection.ltr,
+      Widget? customBody}) {
     _context = context;
-    _progressDialogType = type ?? ProgressDialogType.Normal;
+    _progressDialogType = type;
     _barrierDismissible = isDismissible ?? true;
-    _showLogs = showLogs ?? false;
-    _customBody = customBody ?? null;
-    _direction = textDirection ?? TextDirection.ltr;
+    _showLogs = showLogs;
+    _customBody = (customBody ?? null)!;
+    _direction = textDirection;
     _autoHide = autoHide;
   }
 
   void style({
-    Widget child,
-    double progress,
-    double maxProgress,
-    String message,
-    Widget progressWidget,
-    Color backgroundColor,
-    TextStyle progressTextStyle,
-    TextStyle messageTextStyle,
-    double elevation,
-    TextAlign textAlign,
-    double borderRadius,
-    Curve insetAnimCurve,
-    EdgeInsets padding,
-    Alignment progressWidgetAlignment,
-    Alignment dialogAlignment,
+    Widget? child,
+    double? progress,
+    double? maxProgress,
+    String? message,
+    Widget? progressWidget,
+    Color? backgroundColor,
+    TextStyle? progressTextStyle,
+    TextStyle? messageTextStyle,
+    double? elevation,
+    TextAlign? textAlign,
+    double? borderRadius,
+    Curve? insetAnimCurve,
+    EdgeInsets? padding,
+    Alignment? progressWidgetAlignment,
+    Alignment? dialogAlignment,
   }) {
     if (_isShowing) return;
     if (_progressDialogType == ProgressDialogType.Download) {
@@ -97,12 +95,12 @@ class ProgressDialog {
   }
 
   void update(
-      {double progress,
-      double maxProgress,
-      String message,
-      Widget progressWidget,
-      TextStyle progressTextStyle,
-      TextStyle messageTextStyle}) {
+      {double? progress,
+      double? maxProgress,
+      String? message,
+      Widget? progressWidget,
+      TextStyle? progressTextStyle,
+      TextStyle? messageTextStyle}) {
     if (_progressDialogType == ProgressDialogType.Download) {
       _progress = progress ?? _progress;
     }
@@ -167,10 +165,8 @@ class ProgressDialog {
         if (_showLogs) debugPrint('ProgressDialog shown');
         _isShowing = true;
 
-        if (_autoHide != null) {
-          Future.delayed(_autoHide).then((value) => hide());
-        }
-
+        Future.delayed(_autoHide).then((value) => hide());
+      
         return true;
       } else {
         if (_showLogs) debugPrint("ProgressDialog already shown/showing");
